cleaning up includes, moved simulate_rdme.h to old_c folder
