+++ spatialpy/ssa_sdpd-c-simulation-engine/include/model.h
@@ -8,7 +8,6 @@ See the file LICENSE.txt for details.
 #ifndef model_h
 #define model_h
 
-#include "particle.hpp"
 #include "particle_system.hpp"
 
 namespace Spatialpy{
