+++ spatialpy/ssa_sdpd-c-simulation-engine/include/output.h
@@ -8,7 +8,6 @@ See the file LICENSE.txt for details.
 #ifndef output_h
 #define output_h
 
-#include "particle.hpp"
 #include "particle_system.hpp"
 
 namespace Spatialpy{
